Test move
Fixed error in board - square 27 was the wrong colour
Added the "Thrown" methods for game controller.
Tests on moving of counter all pass
Counter moves the thrown amount and stops moving when it reaches 80. Will only move to 80 with a completing throw of the corrrect number - e.g. if on 75 will only move to 80 (or forward at all) if 5 or less is thrown.